+++ src/core/utils_test.go
@@ -135,19 +135,6 @@ func TestLookPathDoesntExist(t *testing.T) {
 	_, err := LookPath("wibblewobbleflibble", []string{"/usr/local/bin", "/usr/bin", "/bin"})
 	assert.Error(t, err)
 }
-func TestAsyncDeleteDir(t *testing.T) {
-	err := os.MkdirAll("test_dir/a/b/c", DirPermissions)
-	assert.NoError(t, err)
-	err = AsyncDeleteDir("test_dir")
-	assert.NoError(t, err)
-	for i := 0; i < 100; i++ {
-		if !PathExists("test_dir") {
-			return
-		}
-		time.Sleep(100 * time.Millisecond)
-	}
-	assert.False(t, PathExists("test_dir"))
-}
 
 // buildGraph builds a test graph which we use to test IterSources etc.
 func buildGraph() *BuildGraph {
